docs: add simple readme
React scalable is a webpack configured template to clone and code in your scratch projects. This setup supports typescript and have some basic components builded to explore react. Furthermore, this project contains the necessary to run integration and end to end testing with cypress. Maybe, you can send a pr to improve the setup.